+++ svg3d/__init__.py
@@ -1,6 +1,6 @@
 from .shaders import diffuse_shader, hex2rgb, rgb2hex
-from .svg3d import Engine
-from .view import View, _pad_arrays, get_lookat_matrix, get_projection_matrix
+from .svg3d import Engine, Mesh, _pad_arrays
+from .view import View, get_lookat_matrix, get_projection_matrix
 
 __all__ = [
     "_pad_arrays",
